+++ packages/svelte/src/index.ts
@@ -1,2 +1,3 @@
 export { default as Flowchart } from './Flowchart.svelte'
-export type { ConnectParams } from './Flowchart.svelte'
+export type { HandleSide } from '@flowgl/core'
+export { FlowChart, generateId } from '@flowgl/core'