+++ packages/svelte/vite.config.ts
@@ -1,9 +1,9 @@
 import { defineConfig } from 'vite'
-import { svelte } from '@sveltejs/vite-plugin-svelte'
+import { svelte, vitePreprocess } from '@sveltejs/vite-plugin-svelte'
 import { resolve } from 'path'
 
 export default defineConfig({
-  plugins: [svelte({ compilerOptions: { runes: false } })],
+  plugins: [svelte({ preprocess: vitePreprocess() })],
   build: {
     lib: {
       entry:    resolve(__dirname, 'src/index.ts'),